+++ controllers/user/auth-controller.js
@@ -1,5 +1,3 @@
-/* jshint esversion: 6 */
-
 const passport = require('passport');
 
 module.exports = function(data) {
