+++ routers/forum-route.js
@@ -1,5 +1,5 @@
 module.exports = function(app, data) {
-    let controller = require('../../controllers/forum-controller')(data);
+    let controller = require('../controllers/forum/forum-controller')(data);
 
     app
         .get('/forums', controller.listForums)
