+++ routers/home-route.js
@@ -1,7 +1,5 @@
-/* jshint esversion: 6 */
-
 module.exports = function(app, data) {
-    let controller = require('../controllers/home-controller')(data);
+    let controller = require('../controllers/home/home-controller')(data);
 
     app.get('/home', controller.home);
 };
