+++ routers/index.js
@@ -1,3 +1,1 @@
-/* jshint esversion: 6 */
-
 module.exports = require('./routes-loader');
