+++ routers/routes-loader.js
@@ -1,5 +1,3 @@
-/* jshint esversion: 6 */
-
 const fileWalker = require('../utils/file-system-utils').walkDirectorySync;
 
 module.exports = function(app, data) {
